Padnora::Net classes extracted
